replace `Buffer()` with `new Buffer()`
Node v7 currently displays a deprecation warning when using `hyperquest`
due to the use of the `Buffer` constructor without the `new` keyword.

While the wisdom of this decision is being debated and it may very well
be rolled back (at least for releases after v7.1.0), might as well add
the `new` keyword to suppress the warnings in v7.0.0 and the
soon-to-be-released v7.1.0.